+++ src/tdlnm_Cpp.cpp
@@ -17,19 +17,6 @@ using Eigen::MatrixXd;
 using Eigen::VectorXd;
 using Eigen::Lower;
 
-// #ifdef _OPENMP
-//   #include <omp.h>
-//   // [[Rcpp::plugins(openmp)]]
-// #else
-//   #define omp_get_num_threads()  1
-//   #define omp_get_thread_num()   0
-//   #define omp_get_max_threads()  1
-//   #define omp_get_thread_limit() 1
-//   #define omp_get_num_procs()    1
-//   #define omp_set_nested(a)   // empty statement to remove the call
-//   #define omp_get_wtime()        0
-// #endif
-
 /**
  * @brief calculate half of metropolis-hastings ratio for given tree
  *